breakMe V 0.0.1
The very first commit of an evolving experimental learning project.

As I continue down my path, this site will evolve, update, change, and grow. 

The intended use is to provide a place to play around with code and to upload it to share with my friend and mentor Matt Leyrat.